pyproject.toml
[project]
name = "notification-bot"
version = "0.1.0"
description = "Телеграмм бот для уведомлений"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "pydantic-settings[yaml]>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.43",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.14.0",
]

# mypy settings

[tool.mypy]
strict = true
plugins = [
    "pydantic.mypy",
]

# ruff settings

[tool.ruff]
required-version = ">=0.12.1"
src = [
    "bot",
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST", # FastAPI (FAST)
    "YTT", # flake8-2020 (YTT)
    "ANN", # flake8-annotations (ANN)
    # "ASYNC", # flake8-async (ASYNC)
    "S", # flake8-bandit (S)
    "BLE", # flake8-blind-except (BLE)
    "FBT", # flake8-boolean-trap (FBT)
    "B", # flake8-bugbear (B)
    "COM", # flake8-commas (COM)
    "C4", # flake8-comprehensions (C4)
    "DTZ", # flake8-datetimez (DTZ)
    "T10", # flake8-debugger (T10)
    "EM", # flake8-errmsg (EM)
    "ISC", # flake8-implicit-str-concat (ISC)
    "ICN", # flake8-import-conventions (ICN)
    "LOG", # flake8-logging (LOG)
    "G", # flake8-logging-format (G)
    "INP", # flake8-no-pep420 (INP)
    "PIE", # flake8-pie (PIE)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    # "PT", # flake8-pytest-style (PT)
    "Q", # flake8-quotes (Q)
    "RSE", # flake8-raise (RSE)
    "RET", # flake8-return (RET)
    "SLF", # flake8-self (SLF)
    "SIM", # flake8-simplify (SIM)
    "SLOT", # flake8-slots (SLOT)
    "TC", # flake8-type-checking (TC)
    "ARG", # flake8-unused-arguments (ARG)
    "PTH", # flake8-use-pathlib (PTH)
    "FLY", # flynt (FLY)
    "I", # isort (I)
    "C90", # mccabe (C90)
    "N", # pep8-naming (N)
    "PERF", # Perflint (PERF)
    "E", # pycodestyle Error (E)
    "W", # pycodestyle Warning (W)
    # "DOC", # pydoclint (DOC)
    # "D", # pydocstyle (D)
    "F", # Pyflakes (F)
    "PGH", # pygrep-hooks (PGH)
    "PLC", # Pylint Convention (PLC)
    "PLE", # Pylint Error (PLE)
    "PLR", # Pylint Refactor (PLR)
    "PLW", # Pylint Warning (PLW)
    "UP", # pyupgrade (UP)
    "FURB", # refurb (FURB)
    "RUF", # Ruff-specific rules (RUF)
    "TRY", # tryceratops (TRY)
]
ignore = [
    "S101", # S101	assert	Use of assert detected
    "RUF001", # ambiguous-unicode-character-string (RUF001)
    "RUF002", # ambiguous-unicode-character-docstring (RUF002)
    "RUF003", # ambiguous-unicode-character-comment (RUF003)
    "TC006", # runtime-cast-value (TC006)
    "DTZ001", # call-datetime-without-tzinfo (DTZ001)
    "DTZ005", # call-datetime-now-without-tzinfo (DTZ005)
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.extend-per-file-ignores]
"stuff.py" = [
    "T20", # flake8-print (T20)
]
"bot/testing/**.py" = [
    "S311", # S311 Standard pseudo-random generators are not suitable for cryptographic purposes
]
"alembic/**" = [
    "INP001"
]
